requirements.txt
aiogram~=3.10.0
pymongo~=4.8.0
motor~=3.5.0
python-dotenv~=1.0.1
pydantic~=2.7.4
pydantic-settings~=2.3.3